Dataset creation applet complete
In the previous version, data generation was done and saved but the data was not pre-processed. This was useful for statistical exploration of the datasets themselves.

An additional dataset is now saved with all the data, fake and real, shuffled together in a 50/50 ratio as suited for ML treatment. File format of choice has now changed to zipped pickle, since it doesn't alter the Pandas structure and is well compressed.

The notebook is now exportable into a python script which will automatically create new datasets. Added some textual printouts with basic information on the process.